[JSC] Fix ShadowRealm unwinding https://bugs.webkit.org/show_bug.cgi?id=237001
Reviewed by Saam Barati.

JSTests:

* test262/expectations.yaml:

Source/JavaScriptCore:

This patch fixes a crash bug found by test262. Regardless of it is RemoteFunction,
we should handle it as the same way to the other normal host functions except
for setting m_seenRemoteFunction = true flag. Previously, we are early returning,
this is wrong since we should stop unwinding if the caller is entry frame.

* interpreter/Interpreter.cpp:
(JSC::UnwindFunctor::operator() const):

git-svn-id: http://svn.webkit.org/repository/webkit/trunk@290283 268f45cc-cd09-0410-ab3c-d52691b4dbfc