[css-cascade] Support 'revert' in @keyframes https://bugs.webkit.org/show_bug.cgi?id=237073
Reviewed by Darin Adler.

LayoutTests/imported/w3c:

Expect tests to pass.

* web-platform-tests/css/css-cascade/revert-val-006-expected.txt:
* web-platform-tests/css/css-cascade/revert-val-007-expected.txt:
* web-platform-tests/css/css-cascade/revert-val-008-expected.txt:
* web-platform-tests/css/css-cascade/revert-val-010-expected.txt:

Source/WebCore:

Resolver::styleForKeyframe was constructing a RenderStyle with only the
styles specified in the @keyframes. But not taking into account the UA
and user rules that affect the element meant that 'revert' was broken.
Since Builder::applyProperty didn't find the declarations from other
origins, 'revert' just behaved as 'unset'.

This patch solves it by detecting if the @keyframes has a 'revert' value
and in that case it collects the UA and user rules so that it can be
resolved properly.

Tests: imported/w3c/web-platform-tests/css/css-cascade/revert-val-006.html
       imported/w3c/web-platform-tests/css/css-cascade/revert-val-007.html
       imported/w3c/web-platform-tests/css/css-cascade/revert-val-008.html
       imported/w3c/web-platform-tests/css/css-cascade/revert-val-010.html

* style/ElementRuleCollector.cpp:
(WebCore::Style::ElementRuleCollector::addAuthorKeyframeRules):
* style/ElementRuleCollector.h:
* style/StyleResolver.cpp:
(WebCore::Style::Resolver::styleForKeyframe):


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@290457 268f45cc-cd09-0410-ab3c-d52691b4dbfc