operationMaterializeObjectInOSR should DeferTerminationForAWhile instead of just DeferTermination. https://bugs.webkit.org/show_bug.cgi?id=237173 rdar://problem/89450245
Reviewed by Yusuke Suzuki.

JSTests:

* stress/termination-exception-in-operationMaterializeObjectInOSR2.js: Added.

Source/JavaScriptCore:

DeferTermination can throw the TerminationException on destruction of the scope
object.  operationMaterializeObjectInOSR() was using DeferTermination, resulting
in the TerminationException being thrown on exit.  This trips up
validateDFGExceptionHandling testing (which is enabled on Debug builds) because the
TerminationException is unexpected.  The solution is simply to use
DeferTerminationForAWhile instead, and let a more natural trap checkpoint throw
the TerminationException later.

Also refactored notifyDebuggerOfUnwinding() to return early if a debugger is not
present.  This avoids unnecessarily entering and exiting a DeferTermination scope
in notifyDebuggerOfUnwinding().  This was observed while testing this change on
the attached test case.

* ftl/FTLOperations.cpp:
(JSC::FTL::operationMaterializeObjectInOSR):
* interpreter/Interpreter.cpp:
(JSC::notifyDebuggerOfUnwinding):


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@290517 268f45cc-cd09-0410-ab3c-d52691b4dbfc