FTP EPLF does not handle directory https://bugs.webkit.org/show_bug.cgi?id=45113
Patch by Pascal Abresch <nep@packageloss.eu> on 2022-02-28
Reviewed by Darin Adler.

The code used "!result.type" to check for an invalid type. However,
the first type in the FTPEntryType enum is FTPDirectoryEntry, so this
was actually returning true for directories, turning them into "junk"
entries which would not be listed.

Since the code for EPLF can only result in either files, directories,
or uninitialized/junk entries, the check can be removed, returning
result.type is the correct thing to do in all 3 cases.
Source/WebCore:

Test: http/tests/misc/ftp-eplf-directory.py

* loader/FTPDirectoryParser.cpp:
(WebCore::parseOneFTPLine):

LayoutTests:

* http/tests/misc/ftp-eplf-directory-expected.txt: Added.
* http/tests/misc/ftp-eplf-directory.py: Added.
* platform/gtk/TestExpectations:

git-svn-id: http://svn.webkit.org/repository/webkit/trunk@290595 268f45cc-cd09-0410-ab3c-d52691b4dbfc