Compositing/paint invalidation with transforms https://bugs.webkit.org/show_bug.cgi?id=233421 <rdar://problem/85814854>
Reviewed by Alan Bujtas.

Source/WebCore:

When changing transform from something like `translate(0)` to `translate3d(0, 0, 0)
scale(0.5, 0.5)`, we trigger compositing, and thus need to repaint the previous rect of the
layer in its old repaint container. However, the existing code would take the new transform
into account when computing that rect, thus dirtying a rect that is too small.

To fix this, we need to use the cached repaint rects on RenderLayer for this repaint.
However, doing so revealed a bug (via
compositing/shared-backing/repaint-into-shared-backing.html) in how we compute repaint
cached repaint rects in the presence of shared backing. During a compositing update, if a
layer's composited state changed, we'd call computeRepaintRectsIncludingDescendants(), which
computes rect relative to the layer's repaint container. But at this time, the state of
backing sharing in BackingSharingState is in flux (we essentially don't know yet if a layer
can use shared backing). So we may compute and cache repaint rects relative to the wrong
container.

To fix this, leverage the existing m_layersPendingRepaint logic, and delay computing the
repaint rects until we've established our final backing sharing configuration.

Tests: compositing/repaint/composite-and-scale-change.html
       compositing/shared-backing/repaint-into-shared-backing-become-composited.html

* rendering/RenderLayer.cpp:
(WebCore::RenderLayer::setBackingProviderLayer):
(WebCore::RenderLayer::calculateClipRects const):
* rendering/RenderLayerCompositor.cpp:
(WebCore::RenderLayerCompositor::BackingSharingState::issuePendingRepaints):
(WebCore::RenderLayerCompositor::updateBacking):

LayoutTests:

* compositing/repaint/composite-and-scale-change-expected.txt: Added.
* compositing/repaint/composite-and-scale-change.html: Added.
    Test that dumps repaint rects.
* compositing/shared-backing/repaint-into-shared-backing-become-composited-expected.html: Added.
* compositing/shared-backing/repaint-into-shared-backing-become-composited.html: Added.
    Companion to repaint-into-shared-backing.html but toggles compositing in the other
    direction.
* platform/gtk/compositing/repaint/composite-and-scale-change-actual.txt: Added.
* platform/mac-wk1/compositing/repaint/composite-and-scale-change-expected.txt: Added.


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@290628 268f45cc-cd09-0410-ab3c-d52691b4dbfc