[Selection] Selection Range should be clamped by the current value length https://bugs.webkit.org/show_bug.cgi?id=237210
Reviewed by  Darin Adler.

LayoutTests/imported/w3c:

Update the test expectations as some sub-tests are now passing.
* web-platform-tests/html/semantics/forms/textfieldselection/selection-start-end-expected.txt:

Source/WebCore:

As per https://html.spec.whatwg.org/multipage/form-control-infrastructure.html#set-the-selection-range,
if start or end "is greater than the length of the relevant value of the text control, then set it to
the length of the relevant value of the text control".

This CL makes selection range clamp to the current value length.

* html/HTMLTextFormControlElement.cpp:
(WebCore::HTMLTextFormControlElement::setSelectionRange):


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@290635 268f45cc-cd09-0410-ab3c-d52691b4dbfc