[CMake] Disabling ENABLE_WEBCORE is ignored when cmake configuration runs again. https://bugs.webkit.org/show_bug.cgi?id=237170
Reviewed by Fujii Hironori.

CMake variables which has chance to set from outside should be cached in CMake configuration cache.
Unless cacheing, the result of building the generated project is not consisitent because the other
configuration may run while building.

To make it complete, I've changed ENABLE_JAVASCRIPTCORE, ENABLE_WEBCORE and ENABLE_WEBKIT to `option()`
which is stored in cached and reused while building even if the confuguration runs again.

* Source/cmake/OptionsPlayStation.cmake:
* Source/cmake/WebKitCommon.cmake:


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@290636 268f45cc-cd09-0410-ab3c-d52691b4dbfc