Make accessibility/aria-hidden-false-works-in-subtrees.html async to pass in isolated tree mode. https://bugs.webkit.org/show_bug.cgi?id=237051 <rdar://problem/89312156>
Reviewed by Chris Fleizach.

Tools:

* WebKitTestRunner/InjectedBundle/mac/AccessibilityControllerMac.mm:
(WTR::AccessibilityController::accessibleElementById):
Convert the JSStringRef to NSString on the main thread and capture the
NSString to be used on the AX secondary thread. This may be safer since
JSStringRefs are supposed to be used on the main thread.

LayoutTests:

* accessibility/aria-hidden-false-works-in-subtrees.html:
Made both the main script and the iframe onload function async to pass
in isolated tree mode.


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@290673 268f45cc-cd09-0410-ab3c-d52691b4dbfc