[web-animations] changes to font-size should recompute keyframes https://bugs.webkit.org/show_bug.cgi?id=237357
Reviewed by Antti Koivisto.

LayoutTests/imported/w3c:

* web-platform-tests/css/css-animations/responsive/column-width-001-expected.txt:
* web-platform-tests/web-animations/animation-model/keyframe-effects/effect-value-context-expected.txt:
* web-platform-tests/web-animations/animation-model/keyframe-effects/effect-value-context-filling-expected.txt:
* web-platform-tests/web-animations/responsive/fontSize-expected.txt:
* web-platform-tests/web-animations/responsive/perspective-expected.txt:
* web-platform-tests/web-animations/responsive/textIndent-expected.txt:
* web-platform-tests/web-animations/responsive/to-style-change-expected.txt:

Source/WebCore:

When we compute keyframes, it's possible that some values are specified using "em" units
and thus dependent on the current value for font-size. If the font-size changes over time,
we must recompute keyframes to ensure that any such value is updated.

To ensure that we correctly determine when the font-size changed, we pass a pointer to the
style seen when we last udpated animations so that we only consider a change in font-size
provided we've ever resolved animations.

* animation/KeyframeEffect.cpp:
(WebCore::KeyframeEffect::propertyAffectingKeyframeResolutionDidChange):
(WebCore::KeyframeEffect::propertyAffectingLogicalPropertiesDidChange): Deleted.
* animation/KeyframeEffect.h:
* animation/KeyframeEffectStack.cpp:
(WebCore::KeyframeEffectStack::applyKeyframeEffects):
* animation/KeyframeEffectStack.h:
* style/StyleTreeResolver.cpp:
(WebCore::Style::TreeResolver::createAnimatedElementUpdate):
* style/Styleable.h:
(WebCore::Styleable::applyKeyframeEffects const):


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@290730 268f45cc-cd09-0410-ab3c-d52691b4dbfc