Outline-width with transition don't animate correctly https://bugs.webkit.org/show_bug.cgi?id=173708
Reviewed by Antti Koivisto.

Source/WebCore:

Account for a change in outline size when checking whether a style change affects
visual overflow.

No new test for this, I'm not sure how to write one since the issue is failing to
repaint during an animation but WKTR seems to force a repaint for reftests.

* rendering/style/RenderStyle.cpp:
(WebCore::RenderStyle::changeAffectsVisualOverflow const):

LayoutTests:

Add a repaint test that checks that the outline is accounted for during the transition.

* fast/repaint/outline-transition-expected.txt: Added.
* fast/repaint/outline-transition.html: Added.


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@290735 268f45cc-cd09-0410-ab3c-d52691b4dbfc